Refactor parser files
Fix Inotify
